Read models types requested changes
