+++ packages/framework-core/src/decorators/read-model.ts
@@ -1,4 +1,4 @@
-import { Class, ReadModelInterface, RoleAccess } from '@boostercloud/framework-types'
+import { Class, FilterHooks, ReadModelInterface, RoleAccess } from '@boostercloud/framework-types'
 import { Booster } from '../booster'
 import { getPropertiesMetadata } from './metadata'
 
@@ -6,7 +6,7 @@ import { getPropertiesMetadata } from './metadata'
  * Decorator to register a class as a ReadModel
  * @param attributes
  */
-export function ReadModel(attributes: RoleAccess): (readModelClass: Class<ReadModelInterface>) => void {
+export function ReadModel(attributes: RoleAccess & FilterHooks): (readModelClass: Class<ReadModelInterface>) => void {
   return (readModelClass) => {
     Booster.configureCurrentEnv((config): void => {
       if (config.readModels[readModelClass.name]) {
@@ -18,7 +18,7 @@ export function ReadModel(attributes: RoleAccess): (readModelClass: Class<ReadMo
         class: readModelClass,
         properties: getPropertiesMetadata(readModelClass),
         authorizedRoles: attributes.authorize,
-        before: attributes.before,
+        before: attributes.before ?? [],
       }
     })
   }
