+++ packages/framework-integration-tests/src/constants.ts
@@ -1,0 +1,2 @@
+// Both used in Cart entity and read-model-integration.ts
+export const beforeHookProductId = 'before-hook-product-id'
