+++ packages/framework-provider-aws-infrastructure/test/infrastructure/stacks/application-stack.test.ts
@@ -29,6 +29,7 @@ describe('the application stack builder', () => {
       class: readModel,
       authorizedRoles: 'all',
       properties: [],
+      before: [],
     }
   })
   config.env['A_CUSTOM_ENV_VARIABLE'] = 'important-value'
