+++ packages/framework-types/src/concepts/role.ts
@@ -1,7 +1,4 @@
 import { Class } from '../typelevel'
-import { FilterFor } from '../searcher'
-import { UserEnvelope } from '../envelope'
-import { ReadModelInterface } from './read-model'
 
 // eslint-disable-next-line @typescript-eslint/no-empty-interface
 export interface RoleInterface {}
@@ -17,12 +14,6 @@ export interface AuthMetadata {
   readonly skipConfirmation?: boolean
 }
 
-export type BeforeFunction = (
-  filter: FilterFor<Class<ReadModelInterface>>,
-  currentUser?: UserEnvelope
-) => FilterFor<Class<ReadModelInterface>>
-
 export interface RoleAccess {
   readonly authorize: 'all' | Array<Class<RoleInterface>>
-  readonly before?: Array<BeforeFunction>
 }
